event_service/requirements.txt
asgiref==3.8.1
Django==4.2.13
djangorestframework==3.15.2
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.2
redis==3.5.3
gunicorn
django-filter
drf-yasg